feat(swimato-ui): updated header
